Create instruction accounts file
