+++ programs/bpf_loader/benches/serialization.rs
@@ -5,7 +5,7 @@ use {
     solana_pubkey::Pubkey,
     solana_rent::Rent,
     solana_sdk_ids::{bpf_loader, bpf_loader_deprecated},
-    solana_transaction_context::{InstructionAccount, TransactionContext},
+    solana_transaction_context::{instruction_accounts::InstructionAccount, TransactionContext},
 };
 
 fn create_inputs(owner: Pubkey, num_instruction_accounts: usize) -> TransactionContext<'static> {
