+++ programs/loader-v4/src/lib.rs
@@ -27,7 +27,9 @@ use {
     solana_svm_log_collector::{ic_logger_msg, LogCollector},
     solana_svm_measure::measure::Measure,
     solana_svm_type_overrides::sync::Arc,
-    solana_transaction_context::{instruction::InstructionContext, BorrowedInstructionAccount},
+    solana_transaction_context::{
+        instruction::InstructionContext, instruction_accounts::BorrowedInstructionAccount,
+    },
     std::{cell::RefCell, rc::Rc},
 };
